+++ CHANGELOG.md
@@ -44,6 +44,10 @@
   [AndrewDMontgomery](https://github.com/andrewdmontgomery)
   [#4875](https://github.com/realm/SwiftLint/pull/4875)
 
+* Adds `sorted_switch_cases` rule which warns when switch cases are not 
+  alphabetically sorted, and sorts them as correction.  
+  [atacan](https://github.com/atacan)
+
 #### Bug Fixes
 
 * Fix `lower_acl_than_parent` rule rewriter by preserving leading whitespace.